`allow_unknown` flag for tasks lets any arguments be given.
No parsing is done, so the task will have to access the arguments directly through `sys.argv`. This is useful when a task wraps a call to another script and you need to pass the arguments through unmodified. Note that you shouldn't enter any other tasks in the same call to invoke after a task which has `allow_unknown == True` (though you can have other tasks _before_ this). I haven't tested what happens in this case, but consider it undefined.